Simplified priorities and improved token selection semantics.
Priorities now depend only on the terminal type and not also on the length or the input/output parameters of the terminal. Also, priorities must be transitive now.

As a bonus from this, priorities involving non-empty tokens are now always enforced, i.e. the followig situation will never happen: token s has lower priority than token t, s is non-empty or t is non-empty, but both s and t are selected. Note that for two empty tokens s and t it can nevertheless happen that s has lower priority than t, but both s and t are selected.